Time: 183 ms (5.13%) | Memory: 86.8 MB (23.93%) - LeetSync

+++ 1335-maximum-candies-allocated-to-k-children/maximum-candies-allocated-to-k-children.cpp
@@ -1,16 +1,16 @@
 class Solution {
 public:
     int maximumCandies(vector<int>& candies, long long k) {
-        long long answer = 0;
-        long long low = 0, hi = 1e14;
+        int64_t answer = 0;
+        int64_t low = 0, hi = 1e14;
         while(low + 1 < hi){
-            long long mid = low + (hi - low) / 2;
-            long long count = 0;
+            int64_t mid = low + (hi - low) / 2;
+            int64_t count = 0;
             for(int i = 0; i < candies.size(); ++i){
                 count += candies[i] / mid;
             }
             if(count >= k){
-                answer = max<long long>(answer, mid);
+                answer = max<int64_t>(answer, mid);
                 low = mid;
             }
             else{
